requirements.txt
udi_interface>=3.0.18
xmltodict
python-pushover
